# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tuquet Automa Core Bridge API",
    "description": "Lightweight native process launcher, WebSocket event hub, and SQLite storage bridge for Automa WebExtension.",
    "contact": {
      "name": "Tuquet Ecosystem",
      "url": "https://github.com/tuquet/tuquet-automa"
    },
    "license": {
      "name": "MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8765",
      "description": "Local Daemon Bridge Server"
    }
  ],
  "paths": {
...
    "/api/v1/storage/tables": {
      "get": {
        "tags": [
          "Storage"
        ],
        "summary": "List all storage tables",
        "description": "Retrieves schemas for all user data tables stored in SQLite.",
        "operationId": "get_storage_tables",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of tables to return",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "minimum": 0
            },
            "example": 50
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Number of items to skip for pagination (default 0)",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "minimum": 0
            },
            "example": 0
          },
          {
            "name": "search",
            "in": "query",
            "description": "Optional search query to filter tables by name or ID",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of data tables",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StorageTable"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Database read error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Storage"
        ],
        "summary": "Create or update a storage table schema",
        "description": "Defines a new data table schema or modifies column definitions for an existing table.",
        "operationId": "add_storage_table",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StorageTable"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Table schema created or updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StorageTable"
                }
              }
            }
          },
          "400": {
            "description": "Invalid table schema",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Database write error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/storage/tables/{id}": {
      "delete": {
        "tags": [
          "Storage"
        ],
        "summary": "Delete a storage table",
        "description": "Drops the specified data table and purges all of its associated rows from SQLite.",
        "operationId": "delete_storage_table",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Unique table identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Table and rows deleted successfully"
          },
          "404": {
            "description": "Table not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Database deletion error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/storage/tables/{id}/rows": {
      "get": {
        "tags": [
          "Storage"
        ],
        "summary": "Get rows from a storage table",
        "description": "Queries paginated row documents belonging to the specified data table.",
        "operationId": "get_storage_table_rows",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Unique table identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of rows to return (default 50)",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "minimum": 0
            },
            "example": 50
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Number of items to skip for pagination (default 0)",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "minimum": 0
            },
            "example": 0
          },
          {
            "name": "search",
            "in": "query",
            "description": "Optional search query to filter rows by data content",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of table row objects",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TableRow"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Table not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Database query error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Storage"
        ],
        "summary": "Insert a row into a storage table",
        "description": "Appends a new document row into the specified data table.",
        "operationId": "add_storage_table_row",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Unique table identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddTableRowPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Row inserted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddTableRowResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed row data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Table not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Database insertion error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AddTableRowPayload": {
        "allOf": [
          {
            "type": "object",
            "description": "Dynamic column values for the row"
          }
        ],
        "description": "Key-value payload for inserting a new table row",
        "example": {
          "age": 30,
          "name": "John Doe"
        }
      },
      "AddTableRowResponse": {
        "type": "object",
        "description": "Response returned after appending a row to a storage table",
        "required": [
          "id",
          "tableId"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Newly assigned row identifier"
          },
          "tableId": {
            "type": "string",
            "description": "Parent table identifier"
          }
        },
        "example": {
          "id": "row_123",
          "table_id": "tbl_456"
        }
      },
      "ApiErrorResponse": {
        "type": "object",
        "required": [
          "error",
          "status"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
      "StorageTable": {
        "type": "object",
        "description": "User data table schema in Automa storage",
        "properties": {
          "columns": {
            "type": [
              "object",
              "null"
            ],
            "description": "Array of column definitions"
          },
          "columnsIndex": {
            "type": [
              "object",
              "null"
            ],
            "description": "Index mapping for columns"
          },
          "createdAt": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Creation timestamp in milliseconds"
          },
          "id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Unique table identifier"
          },
          "items": {
            "type": [
              "object",
              "null"
            ],
            "description": "Optional inline items cache"
          },
          "modifiedAt": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Last modification timestamp in milliseconds"
          },
          "name": {
            "type": [
              "string",
              "null"
            ],
            "description": "Table display name"
          }
        },
        "example": {
          "id": "tbl_1",
          "name": "Users"
        }
      },
...
      "TableRow": {
        "type": "object",
        "description": "Single data row entry belonging to a storage table",
        "required": [
          "id",
          "table_id",
          "data"
        ],
        "properties": {
          "data": {
            "type": "object",
            "description": "Key-value document payload of row columns"
          },
          "id": {
            "type": "string",
            "description": "Unique row identifier"
          },
          "table_id": {
            "type": "string",
            "description": "Parent table identifier"
          }
        },
        "example": {
          "data": {
            "name": "Alice"
          },
          "id": "row_1",
          "table_id": "tbl_1"
        }
      },
...
    }
  },
...
}